api/requirements.txt
codegen==1.0
Django==2.0.5
django-cors-headers==2.2.0
djangorestframework==3.8.2
drf-nested-routers==0.90.2
psycopg2==2.7.4
pytz==2018.4
uWSGI==2.0.17
